chore: cleanup the number parsing code
